Persist UI choices
This fixes persisting the time type and the sort order in local-storage, similar to dark-mode override.

This fixes #182
